+++ elasticdl/python/common/file_utils.py
@@ -1,7 +1,7 @@
 import os
 import shutil
 
-from elasticdl.python.common.log_util import default_logger as logger
+from elasticdl.python.common.log_utils import default_logger as logger
 
 
 def copy_if_not_exists(src, dst, is_dir):
