+++ elasticdl/python/common/k8s_tensorboard_client.py
@@ -3,7 +3,7 @@ import time
 from kubernetes import client
 
 from elasticdl.python.common import k8s_client as k8s
-from elasticdl.python.common.log_util import default_logger as logger
+from elasticdl.python.common.log_utils import default_logger as logger
 
 
 class TensorBoardClient(object):
