+++ elasticdl/python/common/model_utils.py
@@ -1,7 +1,7 @@
 import importlib.util
 import os
 
-from elasticdl.python.common.log_util import default_logger as logger
+from elasticdl.python.common.log_utils import default_logger as logger
 from elasticdl.python.worker.prediction_outputs_processor import (
     BasePredictionOutputsProcessor,
 )
