+++ elasticdl/python/common/tensor_utils.py
