+++ elasticdl/python/master/checkpoint_service.py
@@ -1,7 +1,7 @@
 import os
 import tempfile
 
-from elasticdl.python.common.model_helper import (
+from elasticdl.python.common.model_utils import (
     load_from_checkpoint_file,
     save_checkpoint_to_file,
 )
