+++ elasticdl/python/master/k8s_worker_manager.py
@@ -3,7 +3,7 @@ import threading
 from collections import Counter
 
 from elasticdl.python.common import k8s_client as k8s
-from elasticdl.python.common.log_util import default_logger as logger
+from elasticdl.python.common.log_utils import default_logger as logger
 
 
 class WorkerManager(object):
